requirements.txt
PySide6==6.9.1
pyinstaller==6.15.0
Pillow==10.4.0
